server/pyproject.toml
[tool.poetry]
name = "meme-generator"
version = "0.1.0"
description = ""
authors = ["stanislavglazko <stanislav.glazko@dualbootpartners.com>"]
readme = "README.md"


[tool.poetry.dependencies]
python = ">=3.10,<3.11"
Django = "^4.2"
djangorestframework = "^3.14.0"
djangorestframework-simplejwt = "^5.2.2"
psycopg2-binary = "^2.9.6"
drf-yasg = "^1.21.7"

[tool.poetry.group.dev.dependencies]
coverage = "^7.2.3"
django-stubs = "^1.16.0"
djangorestframework-stubs = "^1.10.0"
pytest-cov = "^4.0.0"
pytest-django = "^4.5.2"
pytest-xdist = "^3.2.1"
ruff = "^0.0.259"
types-python-dateutil = "^2.8.19.12"
freezegun = "^1.5.1"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "meme_generator.settings"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
